requirements.txt
yfinance
gspread
oauth2client
pandas
numpy
scikit-learn
python-telegram-bot
